32Bit Windows Fix #2

+++ src/src/BranchBound/src/schemes/Branch_Method.jl
@@ -13,9 +13,9 @@ function BM_breadth!(S::BnBSolver,B::BnBModel{T},tL::Float64,tU::Float64,
   unshift!(B.box,X1,X2)
   unshift!(B.LBD,tL,tL)
   unshift!(B.UBD,tU,tU)
-  unshift!(B.id,B.max_id+1,B.max_id+2)
-  unshift!(B.pos,pos+1,pos+1)
-  B.max_id += 2
+  unshift!(B.id,B.max_id+Int64(1),B.max_id+Int64(2))
+  unshift!(B.pos,pos+Int64(1),pos+Int64(1))
+  B.max_id += Int64(2)
 end
 
 """
@@ -33,9 +33,9 @@ function BM_depth_best!(S::BnBSolver,B::BnBModel{T},tL::Float64,tU::Float64,
   push!(B.box,X1,X2)
   push!(B.LBD,tL,tL)
   push!(B.UBD,tU,tU)
-  push!(B.id,B.max_id+1,B.max_id+2)
-  push!(B.pos,pos+1,pos+1)
-  B.max_id += 2
+  push!(B.id,B.max_id+Int64(1),B.max_id+Int64(2))
+  push!(B.pos,pos+Int64(1),pos+Int64(1))
+  B.max_id += Int64(2)
 end
 
 """
